Implement get_next_behaviour to handle choosing the highest priority task, managing task states, and removing completed tasks

+++ src/scheduler/prioritised_task.rs
@@ -2,5 +2,5 @@ use crate::task::Tasks;
 
 pub(super) struct PrioritisedTask {
 	pub(super) task: Tasks,
-	pub(super) priority_function: Box<dyn FnMut() -> u32>,
+	pub(super) priority_function: Box<dyn FnMut() -> usize>,
 }
